docs: reorganize README and add testing.md (#7)
This commit starts to reorganize the README document so that it's more
of a high-level overview with links to other documents and information.
Most of the text from the README had been moved to FAQ.md. The tutorial
has been modified to account for some minor changes in CLI usage. Finally
I have added a testing.md document which can be used to add test cases
for Test Day.

Signed-off-by: Lance Ball <lball@redhat.com>